Added List of Address to a User - @OneToMany relation and endpoints
Create an address model to transform from JSON/XML to java.
Include dependency "modelmapper" to copy the object created with the client information into a UserDto.
Create AddressDto.
Include the list of Address into the UserEntity.
Generate public addres id.
Include list of addresses in the result of the user creation.
Add endpoint to get list of addresses.
Create address service and repository and implement logic.

- Added endpoint for getting all user addresses
- Added endpoint for getting single user address